+++ crates/web-client/test/new_transactions.test.ts
@@ -5,37 +5,66 @@ import {
   mintTransaction,
   setupWalletAndFaucet,
 } from "./webClientTestUtils";
-import { TransactionProver } from "../dist";
 import { setupConsumedNote } from "./notes.test";
 import { Account, TransactionRecord } from "../dist/crates/miden_client_web";
 
 // NEW_MINT_TRANSACTION TESTS
 // =======================================================================================================
 
-describe("new_mint_transactions tests", () => {
-  it("new_mint_transaction completes successfully", async () => {
-    const { faucetId, accountId } = await setupWalletAndFaucet();
-    const result = await mintTransaction(accountId, faucetId);
+describe("mint transaction tests", () => {
+  const testCases = [
+    { flag: false, description: "mint transaction completes successfully" },
+    {
+      flag: true,
+      description: "mint transaction with remote prover completes successfully",
+    },
+  ];
 
-    expect(result.transactionId).to.not.be.empty;
-    expect(result.numOutputNotesCreated).to.equal(1);
-    expect(result.nonce).to.equal("1");
+  testCases.forEach(({ flag, description }) => {
+    it(description, async () => {
+      const { faucetId, accountId } = await setupWalletAndFaucet();
+      const result = await mintTransaction(accountId, faucetId, flag);
+
+      expect(result.transactionId).to.not.be.empty;
+      expect(result.numOutputNotesCreated).to.equal(1);
+      expect(result.nonce).to.equal("1");
+    });
   });
 });
 
 // NEW_CONSUME_TRANSACTION TESTS
 // =======================================================================================================
 
-describe("new_consume_transaction tests", () => {
-  it("new_consume_transaction completes successfully", async () => {
-    const { faucetId, accountId } = await setupWalletAndFaucet();
-    const { createdNoteId } = await mintTransaction(accountId, faucetId);
-    const result = await consumeTransaction(accountId, faucetId, createdNoteId);
+describe("consume transaction tests", () => {
+  const testCases = [
+    { flag: false, description: "consume transaction completes successfully" },
+    {
+      flag: true,
+      description:
+        "consume transaction with remote prover completes successfully",
+    },
+  ];
 
-    expect(result.transactionId).to.not.be.empty;
-    expect(result.nonce).to.equal("1");
-    expect(result.numConsumedNotes).to.equal(1);
-    expect(result.targetAccountBalanace).to.equal("1000");
+  testCases.forEach(({ flag, description }) => {
+    it(description, async () => {
+      const { faucetId, accountId } = await setupWalletAndFaucet();
+      const { createdNoteId } = await mintTransaction(
+        accountId,
+        faucetId,
+        flag
+      );
+      const result = await consumeTransaction(
+        accountId,
+        faucetId,
+        createdNoteId,
+        flag
+      );
+
+      expect(result.transactionId).to.not.be.empty;
+      expect(result.nonce).to.equal("1");
+      expect(result.numConsumedNotes).to.equal(1);
+      expect(result.targetAccountBalanace).to.equal("1000");
+    });
   });
 });
 
@@ -47,8 +76,10 @@ interface SendTransactionResult {
   changedTargetBalance: string;
 }
 
-export const sendTransaction = async (): Promise<SendTransactionResult> => {
-  return await testingPage.evaluate(async () => {
+export const sendTransaction = async (
+  withRemoteProver: boolean
+): Promise<SendTransactionResult> => {
+  return await testingPage.evaluate(async (_withRemoteProver: boolean) => {
     const client = window.client;
 
     const senderAccount = await client.newWallet(
@@ -68,33 +99,67 @@ export const sendTransaction = async (): Promise<SendTransactionResult> => {
     );
     await client.syncState();
 
-    let mintTransactionResult = await client.newMintTransaction(
+    let mintTransactionRequest = client.newMintTransactionRequest(
       senderAccount.id(),
       faucetAccount.id(),
       window.NoteType.private(),
       BigInt(1000)
     );
+    let mintTransactionResult = await client.newTransaction(
+      faucetAccount.id(),
+      mintTransactionRequest
+    );
+    if (_withRemoteProver && window.remoteProverUrl != null) {
+      await client.submitTransaction(
+        mintTransactionResult,
+        window.TransactionProver.newRemoteProver(window.remoteProverUrl)
+      );
+    } else {
+      await client.submitTransaction(mintTransactionResult);
+    }
     let createdNotes = mintTransactionResult.createdNotes().notes();
     let createdNoteIds = createdNotes.map((note) => note.id().toString());
     await window.helpers.waitForTransaction(
       mintTransactionResult.executedTransaction().id().toHex()
     );
 
-    const senderConsumeTransactionResult = await client.newConsumeTransaction(
+    const senderConsumeTransactionRequest =
+      client.newConsumeTransactionRequest(createdNoteIds);
+    let senderConsumeTransactionResult = await client.newTransaction(
       senderAccount.id(),
-      createdNoteIds
+      senderConsumeTransactionRequest
     );
+    if (_withRemoteProver && window.remoteProverUrl != null) {
+      await client.submitTransaction(
+        senderConsumeTransactionResult,
+        window.TransactionProver.newRemoteProver(window.remoteProverUrl)
+      );
+    } else {
+      await client.submitTransaction(senderConsumeTransactionResult);
+    }
     await window.helpers.waitForTransaction(
       senderConsumeTransactionResult.executedTransaction().id().toHex()
     );
 
-    let sendTransactionResult = await client.newSendTransaction(
+    let sendTransactionRequest = client.newSendTransactionRequest(
       senderAccount.id(),
       targetAccount.id(),
       faucetAccount.id(),
       window.NoteType.private(),
       BigInt(100)
     );
+    let sendTransactionResult = await client.newTransaction(
+      senderAccount.id(),
+      sendTransactionRequest
+    );
+    if (_withRemoteProver && window.remoteProverUrl != null) {
+      await client.submitTransaction(
+        sendTransactionResult,
+        window.TransactionProver.newRemoteProver(window.remoteProverUrl)
+      );
+    } else {
+      await client.submitTransaction(sendTransactionResult);
+    }
     let sendCreatedNotes = sendTransactionResult.createdNotes().notes();
     let sendCreatedNoteIds = sendCreatedNotes.map((note) =>
       note.id().toString()
@@ -103,10 +168,20 @@ export const sendTransaction = async (): Promise<SendTransactionResult> => {
       sendTransactionResult.executedTransaction().id().toHex()
     );
 
-    const targetConsumeTransactionResult = await client.newConsumeTransaction(
+    const targetConsumeTransactionRequest =
+      client.newConsumeTransactionRequest(sendCreatedNoteIds);
+    let targetConsumeTransactionResult = await client.newTransaction(
       targetAccount.id(),
-      sendCreatedNoteIds
+      targetConsumeTransactionRequest
     );
+    if (_withRemoteProver && window.remoteProverUrl != null) {
+      await client.submitTransaction(
+        targetConsumeTransactionResult,
+        window.TransactionProver.newRemoteProver(window.remoteProverUrl)
+      );
+    } else {
+      await client.submitTransaction(targetConsumeTransactionResult);
+    }
     await window.helpers.waitForTransaction(
       targetConsumeTransactionResult.executedTransaction().id().toHex()
     );
@@ -124,15 +199,25 @@ export const sendTransaction = async (): Promise<SendTransactionResult> => {
         .getBalance(faucetAccount.id())
         .toString(),
     };
-  });
+  }, withRemoteProver);
 };
 
-describe("new_send_transaction tests", () => {
-  it("new_send_transaction completes successfully", async () => {
-    const result = await sendTransaction();
+describe("send transaction tests", () => {
+  const testCases = [
+    { flag: false, description: "send transaction completes successfully" },
+    {
+      flag: true,
+      description: "send transaction with remote prover completes successfully",
+    },
+  ];
 
-    expect(result.senderAccountBalance).to.equal("900");
-    expect(result.changedTargetBalance).to.equal("100");
+  testCases.forEach(({ flag, description }) => {
+    it(description, async () => {
+      const result = await sendTransaction(flag);
+
+      expect(result.senderAccountBalance).to.equal("900");
+      expect(result.changedTargetBalance).to.equal("100");
+    });
   });
 });
 
@@ -141,10 +226,10 @@ describe("new_send_transaction tests", () => {
 
 export const customTransaction = async (
   assertedValue: string,
-  withCustomProver: boolean
+  withRemoteProver: boolean
 ): Promise<void> => {
   return await testingPage.evaluate(
-    async (_assertedValue: string, _withCustomProver: boolean) => {
+    async (_assertedValue: string, _withRemoteProver: boolean) => {
       const client = window.client;
 
       const walletAccount = await client.newWallet(
@@ -353,8 +438,11 @@ export const customTransaction = async (
         transactionRequest
       );
 
-      if (_withCustomProver) {
-        await client.submitTransaction(transactionResult, await selectProver());
+      if (_withRemoteProver && window.remoteProverUrl != null) {
+        await client.submitTransaction(
+          transactionResult,
+          window.TransactionProver.newRemoteProver(window.remoteProverUrl)
+        );
       } else {
         await client.submitTransaction(transactionResult);
       }
@@ -405,10 +493,10 @@ export const customTransaction = async (
         transactionRequest2
       );
 
-      if (_withCustomProver) {
+      if (_withRemoteProver && window.remoteProverUrl != null) {
         await client.submitTransaction(
           transactionResult2,
-          await selectProver()
+          window.TransactionProver.newRemoteProver(window.remoteProverUrl)
         );
       } else {
         await client.submitTransaction(transactionResult2);
@@ -419,7 +507,7 @@ export const customTransaction = async (
       );
     },
     assertedValue,
-    withCustomProver
+    withRemoteProver
   );
 };
 
@@ -526,6 +614,10 @@ describe("custom transaction tests", () => {
   it("custom transaction fails", async () => {
     await expect(customTransaction("1", false)).to.be.rejected;
   });
+
+  it("custom transaction with remote prover completes successfully", async () => {
+    await expect(customTransaction("0", true)).to.be.fulfilled;
+  });
 });
 
 describe("custom transaction with multiple output notes", () => {
@@ -552,24 +644,6 @@ describe("custom transaction with multiple output notes", () => {
       }
     });
   });
-});
-
-// CUSTOM PROVERS TEST
-// ================================================================================================
-
-export const selectProver = async (): Promise<TransactionProver> => {
-  if (window.remoteProverUrl != null) {
-    return window.TransactionProver.newRemoteProver(window.remoteProverUrl);
-  } else {
-    return window.TransactionProver.newLocalProver();
-  }
-};
-
-describe("use custom transaction prover per request", () => {
-  it("custom transaction prover completes successfully"),
-    async () => {
-      await expect(customTransaction("0", true)).to.be.fulfilled;
-    };
 });
 
 // DISCARDED TRANSACTIONS TESTS
@@ -604,27 +678,35 @@ export const discardedTransaction =
       );
       await client.syncState();
 
-      let mintTransactionResult = await client.newMintTransaction(
+      let mintTransactionRequest = client.newMintTransactionRequest(
         senderAccount.id(),
         faucetAccount.id(),
         window.NoteType.private(),
         BigInt(1000)
       );
+      let mintTransactionResult = await client.newTransaction(
+        faucetAccount.id(),
+        mintTransactionRequest
+      );
+      await client.submitTransaction(mintTransactionResult);
       let createdNotes = mintTransactionResult.createdNotes().notes();
       let createdNoteIds = createdNotes.map((note) => note.id().toString());
       await window.helpers.waitForTransaction(
         mintTransactionResult.executedTransaction().id().toHex()
       );
 
-      const senderConsumeTransactionResult = await client.newConsumeTransaction(
+      const senderConsumeTransactionRequest =
+        client.newConsumeTransactionRequest(createdNoteIds);
+      let senderConsumeTransactionResult = await client.newTransaction(
         senderAccount.id(),
-        createdNoteIds
+        senderConsumeTransactionRequest
       );
+      await client.submitTransaction(senderConsumeTransactionResult);
       await window.helpers.waitForTransaction(
         senderConsumeTransactionResult.executedTransaction().id().toHex()
       );
 
-      let sendTransactionResult = await client.newSendTransaction(
+      let sendTransactionRequest = client.newSendTransactionRequest(
         senderAccount.id(),
         targetAccount.id(),
         faucetAccount.id(),
@@ -632,6 +714,11 @@ export const discardedTransaction =
         BigInt(100),
         1
       );
+      let sendTransactionResult = await client.newTransaction(
+        senderAccount.id(),
+        sendTransactionRequest
+      );
+      await client.submitTransaction(sendTransactionResult);
       let sendCreatedNotes = sendTransactionResult.createdNotes().notes();
       let sendCreatedNoteIds = sendCreatedNotes.map((note) =>
         note.id().toString()
@@ -653,11 +740,13 @@ export const discardedTransaction =
       let preConsumeStore = await client.exportStore();
 
       // Sender retrieves the note
-      let senderTxResult = await client.newConsumeTransaction(
+      let senderTxRequest =
+        await client.newConsumeTransactionRequest(sendCreatedNoteIds);
+      let senderTxResult = await client.newTransaction(
         senderAccount.id(),
-        sendCreatedNoteIds
+        senderTxRequest
       );
-
+      await client.submitTransaction(senderTxResult);
       await window.helpers.waitForTransaction(
         senderTxResult.executedTransaction().id().toHex()
       );
@@ -722,6 +811,32 @@ export const discardedTransaction =
       };
     });
   };
+
+describe("custom transaction with multiple output notes", () => {
+  const testCases = [
+    {
+      description: "does not fail when output note serial numbers are unique",
+      shouldFail: false,
+    },
+    {
+      description: "fails when output note serial numbers are the same",
+      shouldFail: true,
+    },
+  ];
+
+  testCases.forEach(({ description, shouldFail }) => {
+    it(description, async () => {
+      const { accountId, faucetId } = await setupConsumedNote();
+      if (shouldFail) {
+        await expect(customTxWithMultipleNotes(shouldFail, accountId, faucetId))
+          .to.be.rejected;
+      } else {
+        await expect(customTxWithMultipleNotes(shouldFail, accountId, faucetId))
+          .to.be.fulfilled;
+      }
+    });
+  });
+});
 
 describe("discarded_transaction tests", () => {
   it("transaction gets discarded", async () => {