+++ crates/web-client/test/transactions.test.ts
@@ -61,6 +61,7 @@ describe("get_transactions tests", () => {
     const { transactionId: mintTransactionId } = await mintTransaction(
       accountId,
       faucetId,
+      false,
       false
     );